Refactor codebase
Splitted up the code in the `index.js` file into 3 separate files for each OS.
