+++ lib/windows.js
@@ -32,7 +32,7 @@ const kernel32 = new ffi.Library('kernel32', {
 	QueryFullProcessImageNameW: ['int', ['pointer', 'uint32', 'pointer', 'pointer']]
 });
 
-module.exports = () => {
+function windows() {
 	// Windows C++ APIs' functions are declared with capitals, so this rule has to be turned off
 	/* eslint-disable new-cap */
 
@@ -86,4 +86,7 @@ module.exports = () => {
 		app: processName,
 		pid: processId
 	};
-};
+}
+
+module.exports = () => Promise.resolve(windows());
+module.exports.sync = windows;
